requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
Django==4.2.20
django-axes==8.0.0
django-debug-toolbar==5.2.0
django-ipware==7.0.1
mypy-extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
python-ipware==3.0.0
sqlparse==0.5.3
tomli==2.2.1
typing-extensions==4.13.2
